Estado conectado o desconectado de los usuarios

+++ frontend/src/auth/logout/index.js
@@ -5,11 +5,29 @@ import "../../static/css/auth/authPage.css";
 import tokenService from "../../services/token.service";
 
 const Logout = () => {
+
+  const jwt = tokenService.getLocalAccessToken();
+
+
   function sendLogoutRequest() {
-    const jwt = window.localStorage.getItem("jwt");
     if (jwt || typeof jwt === "undefined") {
-      tokenService.removeUser();
-      window.location.href = "/";
+
+      fetch("/api/v1/profile/disconnect", {
+        method: "PATCH",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${jwt}`,
+        },
+      }).then((response) => {
+        if (response.ok) {
+          tokenService.removeUser();
+          window.localStorage.removeItem("jwt");
+          window.location.href = "/";
+        } else {
+          alert("There was an error logging out");
+        }
+      }
+      );
     } else {
       alert("There is no user logged in");
     }
@@ -22,7 +40,7 @@ const Logout = () => {
           Are you sure you want to log out?
         </h2>
         <div className="options-row">
-          <Link className="auth-button" to="/home" style={{textDecoration: "none"}}>
+          <Link className="auth-button" to="/home" style={{ textDecoration: "none" }}>
             No
           </Link>
           <button className="auth-button" onClick={() => sendLogoutRequest()}>
